backend/requirements_flask.txt
# Flask API for website (behavior log & analytics & dashboard)
# Install: pip install -r requirements_flask.txt

flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
PyJWT>=2.8.0
werkzeug>=3.0.0
flasgger>=0.9.7
